requirements.txt
streamlit>=1.36
httpx>=0.27
python-dotenv>=1.0
pandas>=2.0
Pillow>=10.0
